Switch to using environment variables for configuration (#6)
Switch from using command line flags to environment variables for
configuration of this service. This is done using a Configuration class
that holds all values as members. A factory function is provided that
pulls the values from environment variables. Code that needs this
configuration takes it in as a constructor argument. This is to make
automated testing easy in the future.

Also switched off of using the Gradle Wrapper. We don't really have much
of a use for it considering this project is intended to be built and run
in Docker.